1bit: use double for scaling (#22)
* 1bit: use double

* 1bit: fix

* misc: new line eof

+++ byteps/common/compressor/strategy/onebit.cc
@@ -49,7 +49,7 @@ size_t OnebitCompressor::PackingImpl(index_t* dst, const scalar_t* src,
 
   float scale = 1.0f;
   if (_use_scale) {
-    float sum = 0.0f;
+    double sum = 0.0f;
     for (size_t i = 0; i < len; ++i) {
       dst[i] = src[i] < 0;
       sum += abs(src[i]);
